pyproject.toml
[tool.poetry]
name = "moncap-api"
version = "0.1.0"
description = "Backend MONCAP (FastAPI + PostgreSQL + JWT)"
authors = ["MONCAP <dev@moncap.local>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.110"
uvicorn = {version = ">=0.27", extras = ["standard"]}
sqlalchemy = {version = ">=2.0", extras = ["asyncio"]}
asyncpg = ">=0.29"
alembic = ">=1.13"
pydantic-settings = ">=2.2"
python-jose = {version = ">=3.3", extras = ["cryptography"]}
passlib = {version = ">=1.7", extras = ["argon2"]}
argon2-cffi = ">=23.1"
python-multipart = ">=0.0.9"
email-validator = ">=2.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
httpx = ">=0.27"
aiosqlite = ">=0.20"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
